fix(telemetry): emit int-typed registry attributes as i64 (#1013)
The tracing→OTel bridge visitors implement record_i64 but not
record_u64, so u64/usize span and event fields fall through to
record_debug and export as strings — violating the registry's
`type: int` declarations under weaver live-check (surfaced on #1007:
signaldb.job.{tables_processed,partitions_dropped,snapshots_expired,
bytes_reclaimed,duration_ms} arrived as strings once the monolith
started running retention in the check window).

Cast every int-typed registry attribute emission to i64 at the call
site (retention enforcer events, WAL and writer span fields), and pin
the rule in registry_pins: any `type: int` attribute assigned without
an `as i64` cast fails the drift gate.

+++ src/writer/src/processor.rs
@@ -449,7 +449,7 @@ impl WalProcessor {
             signaldb.tenant.id = %tenant_id,
             signaldb.dataset.id = %dataset_id,
             signaldb.table = %table_name,
-            signaldb.wal.entry_count = entries.len()
+            signaldb.wal.entry_count = entries.len() as i64
         )
     )]
     async fn process_batch_for_table(
